Rename `beacon/main.py` -> `beacon/beacon.py`
