+++ web3/beacon/__init__.py
@@ -1,2 +1,2 @@
 from .async_beacon import AsyncBeacon
-from .main import Beacon
+from .beacon import Beacon
