+++ web3/beacon/beacon.py
